search-api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
elasticsearch==8.11.0
qdrant-client==1.7.0
sentence-transformers>=2.3.0
torch>=2.0.0
numpy==1.24.3
pyyaml==6.0.1
python-multipart==0.0.6
loguru==0.7.2
tenacity==8.2.3
cachetools==5.3.2

